Add default environment prefix option
The reason for the massive meta lookup change is because environment
variables could be retrieved before the meta ever gets defined. We
could simply require users to define the meta first but that would be
a confusing API. A downside of this change is that the meta gets
executed twice which might lead to some problems.