Make default settings accessible directly from the class
The __class__ assignment was removed because trying to get a setting
from the Settings class recursed infinitely due to the Settings class
not having a _meta attribute. Other possible fixes would've been making
a special case for Settings or directly assigning the _meta attribute
ourselves but removing the __class__ assignment seemed to be the cleaner
alternative. If this causes any issues in the future changing the
implementation to one of the aforementioned should be considered.