Fix compile breaking on Python 3.8
CPython's merge with typed_ast changed the AST, including adding a
type_ignores field to ast.Module which this commit now applies.

Something interesting to note is that due to how the init for AST nodes
work you can pass in arbitrary keyword arguments, unlike positional
arguments which strictly map to the nodes' fields. Because of this, we
can keep compatibility with older Python versions by using keyword
arguments for newly added fields. The implementation for it is here:
https://github.com/python/cpython/blob/235e7b2b0d937ba8b4a9172aa72206187e3e1f54/Python/Python-ast.c#L560