Use a single job for testing
Instead of trying to integrate tox into travis and creating a disconnect
between local test runs and CI ones, we just install the required Python
versions ourselves and invoke the test runs normally in a single job. As
a bonus, we get to keep the coverage report hidden in a drop-down.

The biggest downside of relying on tox is that ignoring the outcomes of
environments makes them useless when running in parallel. Parallel runs
only have binary statuses so if an ignored run fails the status of it is
visually the same as it succeeding.

This also fixes the lint job accidentally reporting code coverage.