Test against installed versions of the library
`poetry install` replaces the installed version with a `.egg-link` that
points to the source. Passing the newly added `--no-root` would fix
this, but if we wanted to test against specific versions of Django it
would clobber them with newer versions. Exporting the dependencies via
`poetry export` and then installing them directly with pip also wouldn't
work as pip doesn't have an `--upgrade-strategy` that completely
disables upgrades. Plus, we would then need to keep our lock file up to
date and since we can't change where it's located test runs would modify
it, defeating its purpose. We could switch to using nox -- which would
solve all our problems -- but then we'd lose the benefit of parallelism.
And so, we just duplicate dependencies only needed for the test run,
which does provide a speed boost.